Change termination keybinding to Ctrl+Alt+Backspace
Prior to this commit, we were using the function
keyboard_execute_compositor_binding() to terminate the compositor when
the special key XKB_KEY_XF86Display was pressed. The main problem with
this approach is that, as stated at the comment describing the function,
it can only handle single-key bindings.
Since we have already changed cmd_bindsym to bind to callbacks instead
of sway commands, we provide a callback to call sway_terminate() and
bind it in a similar fashion to the Alt+Enter keybinding used to spawn
a terminal. This allows us to have a more natural multi-key binding
to terminate the server.