apply_horiz_layout(): use same width fraction for all children
The previous method for calculating children's width used the fraction
of width they previously occupied, as a remnant of containers' former
tree-like structure. This made it so that, if we had output O1 with two
containers (A and B) occupying half screen-width each, and output 2
with one container (C), then unplugging output O2
(and consequently transferring container C to output O1)
would result in containers A and B having 25% of O1's total screen
width each, while C would have the remaining 50%, instead of each
container occupying 1/3 of the screen width, as will be the case due
to this commit.