Refactor (almost inline) function output_add_workspace()
This function is only called at workspace_create(). It is almost inlined
at its call site, except the condition
```
if (workspace->output) {
        ...
}
```
Was ommited entirely since a newly created workspace can't have an
output assigned to it yet. Also, the other `if` check at
output_add_workspace() was moved to the beginning of
workspace_create().