Replace calls to seat_get_focus_inactive() using root node
Replace all calls to
        seat_get_focus_inactive(<seat pointer>, &root->node);
with a new function seat_get_next_in_focus_stack(), in an effort
to reduce uses of the root node (so we can remove it later.)

Analysing seat_get_focus_inactive(), the condition

    if (node_is_view(node)) {
        return node;
    }

is necessarily false when `node` is `root->node`, and

    struct sway_seat_node *current;
    wl_list_for_each(current, &seat->focus_stack, link) {
        if (node_has_ancestor(current->node, node)) {
            return current->node;
        }
    }

will return the first node whose ancestor can be traced back to the
root node, which happens to all nodes _except_ for detached workspaces,
or containers without an output. The latter never seems to happen, so,
in the new function seat_get_next_in_focus_stack(), we simply return
the first node in the focus stack which is not a detached workspace.

Additionally, assert that the new function returns the same as
seat_get_focus_inactive() would if called with the root node.