Move sway_container code that relies on sway_seat to separate file
Make a new file mod_sway/conainer.c and place there all
container-related functions that rely on sway_seat and other types
on the input folder.
The main reason for this change is the upcoming migration of
sway_container to the wlstem library, which will not *yet*
receive input-related code; that's expected to happen in the future.