Move (most) title data away from sway_container
Since sway_container's definition is already in wlstem,
remove some its fields which do not belong in this library, namely,
`formatted_title` and the textures `title_focused`, `title_unfocused`
and `title_urgent`.
The decision of how to format title, along with which and how many
textures to use for title and when to use each one are all
implementation decisions to be made not by this library, but by the
programs using it.