HACK: comment out input_manager_current_seat call
We wish to migrate transactions to wlstem, but as it is now,
copy_container_state() calls input_manager_current_seat(), a function
that depends on several others, many of which still have WM code. As
such, break the dependency on this function by commenting out the call
and hardcoding `state->focused` as false, which has the side effect that
all window titles are rendered with an "unfocused" (gray) color.