output_render(): Refactor most actual rendering into seperate functions
output_render() has some bookkeeping code that we wish to migrate to
wlstem, but most of the actual rendering code is considered WM-specific
and as such, this code should be given by library users.
Therefore, this code into two separate functions, output_render_overlay
and output_render_non_overlay, which shall be passed as callbacks to
wlstem later on.