output_evacuate(): move choice of fallback output to new function
When an output is disconnected, the windows (sway_containers) it had
must be transferred to a different output (possibly including
noop_output). This choice is considered a window management one, but it
must to be done, or else we would have a dangling pointer in
sway_container's `output` field. As such, migrate the code that makes
this choice, in output_evacuate(), to a separate function
choose_absorber_output(), so that it can be provided as a callback to
wlstem later.