meson.build: Add explicit `check: false` in run_command calls
The boolean argument `check` of `run_command` determines whether the
build should fail in case the command invoked by `run_command` fails.
Since we're only using this to determine a custom version from commit
and branch, we don't want the build to fail if these commands fail.
So far, the default value for `check` was false, but this might change
soon, so use an explicit `check: false` in `run_command`.